refactor interactive module
